release/requirements.txt
customtkinter==5.2.1
Pillow==10.1.0
plyer==2.1.0
pywin32==306
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
reportlab==4.0.7
tkcalendar==1.6.1 